third_party/WebKit/LayoutTests/fast/css/parsing-css-string-characters.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../css-parser/resources/property-parsing-test.js"></script>
<script>
assert_valid_value("font-family", '"65: A 9: \t"', '"65: A 9: \\9 "');
assert_valid_value("font-family", '"128: \u0080 159: \u009f"');
assert_valid_value("font-family", '"1: \x01"', '"1: \\1 "');
assert_valid_value("font-family", '"8: \x08"', '"8: \\8 "');
assert_valid_value("font-family", '"9: \t"', '"9: \\9 "');
assert_valid_value("font-family", '"11: \x0b"', '"11: \\b "');
assert_valid_value("font-family", '"14: \x0e"', '"14: \\e "');
assert_valid_value("font-family", '"20: \x14"', '"20: \\14 "');
assert_valid_value("font-family", '"30: \x1e"', '"30: \\1e "');
assert_valid_value("font-family", '"31: \x1f"', '"31: \\1f "');
assert_valid_value("font-family", '"127: \x7f"', '"127: \\7f "');
assert_valid_value("font-family", '"384: \u0180"');

assert_invalid_value("font-family", '"10: \n"');
assert_invalid_value("font-family", '"12: \f"');
assert_invalid_value("font-family", '"13: \r"');
</script>
